fix(phase6): serialize Decimal score fields as float in approval schemas
Pydantic v2 serializes Decimal as a JSON string by default. The
ManagerApprovals page calls .toFixed() on overall_score and raw_score,
which throws "TypeError: e.toFixed is not a function" when the value
is a string, crashing the React render tree and leaving a blank page.

Change overall_score in ApprovalQueueItemOut and SubmissionOut, and
raw_score in ComponentScoreOut, from Decimal to float so they serialize
as JSON numbers. The Decimal import is retained for the input schemas
(ApprovalDecisionRequest, approved_purchase_offer, etc.) where financial
precision on inbound values still matters.

+++ api/schemas/appraisal_submission.py
@@ -56,7 +56,7 @@ class ComponentScoreOut(BaseModel):
     id: uuid.UUID
     component_id: uuid.UUID
     component_name: str
-    raw_score: Decimal
+    raw_score: float
     weight_at_time_of_scoring: Decimal
     notes: str | None
 
@@ -77,7 +77,7 @@ class SubmissionOut(BaseModel):
     running_status: str | None
     serial_number: str | None
     title_status: str | None
-    overall_score: Decimal | None
+    overall_score: float | None
     score_band: str | None
     management_review_required: bool
     hold_for_title_review: bool
